Move in separate files the logic to recreate the images for regression tests
